+++ app/src/main/java/com/tienda/a_shop/presenters/interfaces/presenters/ICategoriaPresenter.java
@@ -7,7 +7,7 @@ package com.tienda.a_shop.presenters.interfaces.presenters;
 
 public interface ICategoriaPresenter {
 
-    void actualizarCategoría(String nombreN, String nombre, int estimado, int orden);
+    void actualizarCategoría(String nombreN, String nombre, double estimado, int orden);
     void agregarCategoria(String nombre, int estimado, int orden);
     void eliminarCategoria(String nomCategoria);
     void listarCategoriasMesActual();
